Made an AudioOutput class, which currently outputs an 100Hz sine wave as PCM data. Seems to work fine, except for volume control. Changing the amplitude of the sine wave does not seem to have any effect on the output volume, so maybe the audio device renormalizes the samples. Have to look into that

+++ libdmg/emulator.cpp
@@ -58,7 +58,7 @@ void Emulator::Step()
 
 
 	// Delay next CPU instruction until we've caught up 
-	ticksUntilNextInstruction += (cpu.Ticks() - previousTicks);
+	ticksUntilNextInstruction += (uint32_t) (cpu.Ticks() - previousTicks);
 
 	// Update IO subsystems
 	video.Sync(ticks);
